Refactor code structure for improved readability and maintainability

+++ generador.c
@@ -2,35 +2,45 @@
 #include <stdlib.h>
 #include <time.h>
 
-int main() {
+int main()
+{
     FILE *f = fopen("medidas.txt", "w");
-    if (!f) {
+    if (!f)
+    {
         perror("Error creando medidas.txt");
         return 1;
     }
 
     srand(time(NULL));
 
-    int activo = 1;   // 1 = sistema encendido, 0 = apagado
-    float temp = 45.0;
+    int activo = 1; // 1 = sistema encendido, 0 = apagado
+    int temp = 45;
 
-    for (int i = 0; i < 40; i++) {
-        if (activo) {
+    for (int i = 0; i < 40; i++)
+    {
+        if (activo)
+        {
             // Aumenta la temperatura en saltos grandes
             temp += (rand() % 40) + 10; // +10 a +50
-            if (temp > 105) temp = 105;
-        } else {
+            if (temp > 105)
+                temp = 105;
+        }
+        else
+        {
             // Disminuye en saltos grandes
             temp -= (rand() % 40) + 10; // -10 a -50
-            if (temp < 45) temp = 45;
+            if (temp < 45)
+                temp = 45;
         }
 
         // Escribir temperatura al archivo
-        fprintf(f, "%.2f\n", temp);
+        fprintf(f, "%d\n", temp);
 
         // Lógica de activación / desactivación
-        if (temp > 90) activo = 0;   // Se apaga al superar 90
-        if (temp < 55) activo = 1;   // Se enciende al bajar de 55
+        if (temp > 90)
+            activo = 0; // Se apaga al superar 90
+        if (temp < 55)
+            activo = 1; // Se enciende al bajar de 55
     }
 
     fclose(f);
